Added another table and it's CRUD
Added CRUD functionalities to Admin table. I believe that this is the end of the backed side of the project. Unless there is something else to add that is not working well. But for now, I believe that the project's implementation is complete. Now, I am going to focus on the graphical user interface.